Added new fightCalculations file to hold all calulcations
